Move JSPIContext out of VM.h https://bugs.webkit.org/show_bug.cgi?id=309295 rdar://170155664
Reviewed by Keith Miller.

This addresses a polish to-do item created when first landing JSPI support. For better
code organization, struct JSPIContext is moved from VM.h into its own dedicated header
file.

To avoid circular includes this would otherwise create or forward declarations making
SaferCPP unhappy, struct PinballCompletionContext is also moved from PinballCompletion.h
into its own header.

Actual logic is unchanged.

Testing: covered by existing tests.
Canonical link: https://commits.webkit.org/308814@main
